chore: removed unused methods in the object storage classes
