+++ posthog/storage/recordings/__init__.py
@@ -1,0 +1,1 @@
+# Storage modules for session recordings
