--- posthog/storage/session_recording_v2_object_storage.py
@@ -1,913 +1,0 @@
-import re
-import abc
-from collections.abc import AsyncIterator
-from contextlib import asynccontextmanager
-from typing import Optional, Protocol, runtime_checkable
-from urllib.parse import parse_qs, urlparse
-
-from django.conf import settings
-
-import snappy
-import aiohttp
-import aioboto3
-import structlog
-import posthoganalytics
-from boto3 import client as boto3_client
-from botocore.client import Config
-
-logger = structlog.get_logger(__name__)
-
-
-class BlockDeleteError(Exception):
-    pass
-
-
-class BlockFetchError(Exception):
-    pass
-
-
-class FileFetchError(Exception):
-    pass
-
-
-class FileDownloadError(Exception):
-    pass
-
-
-class FileUploadError(Exception):
-    pass
-
-
-class FileDeleteError(Exception):
-    pass
-
-
-class RecordingApiFetchError(Exception):
-    pass
-
-
-class RecordingDeletedError(Exception):
-    """Raised when attempting to access a recording that has been deleted (crypto shredding)."""
-
-    def __init__(self, message: str, deleted_at: Optional[int] = None):
-        super().__init__(message)
-        self.deleted_at = deleted_at
-
-
-@runtime_checkable
-class BlockStorage(Protocol):
-    """
-    Protocol for fetching recording blocks.
-
-    Both the S3 storage client and the Recording API client implement this interface,
-    allowing them to be used interchangeably when fetching blocks.
-
-    The session_id and team_id parameters are required for the Recording API client
-    to construct the correct URL, but are ignored by the S3 storage client.
-    """
-
-    async def fetch_block(self, block_url: str, session_id: str, team_id: int) -> str:
-        """Fetch and decompress a recording block. Returns decompressed content."""
-        ...
-
-    async def fetch_block_bytes(self, block_url: str, session_id: str, team_id: int) -> bytes:
-        """Fetch a recording block. Returns compressed bytes."""
-        ...
-
-
-class SessionRecordingV2ObjectStorageBase(metaclass=abc.ABCMeta):
-    @abc.abstractmethod
-    def read_bytes(self, key: str, first_byte: int, last_byte: int) -> bytes | None:
-        pass
-
-    @abc.abstractmethod
-    def read_all_bytes(self, key: str) -> bytes | None:
-        pass
-
-    @abc.abstractmethod
-    def write(self, key: str, data: bytes) -> None:
-        pass
-
-    @abc.abstractmethod
-    def is_enabled(self) -> bool:
-        pass
-
-    @abc.abstractmethod
-    def fetch_file(self, blob_key: str) -> str:
-        """Returns the decompressed file or raises Error"""
-        pass
-
-    @abc.abstractmethod
-    def fetch_file_bytes(self, blob_key: str) -> bytes:
-        """Returns the compressed file as bytes or raises FileFetchError"""
-        pass
-
-    @abc.abstractmethod
-    def fetch_block(self, block_url: str) -> str:
-        """Returns the decompressed block or raises BlockFetchError"""
-        pass
-
-    @abc.abstractmethod
-    def fetch_block_bytes(self, block_url: str) -> bytes:
-        """Returns the compressed block as bytes or raises BlockFetchError"""
-        pass
-
-    @abc.abstractmethod
-    def delete_block(self, block_url: str) -> None:
-        """Zeroes out the specified block or raises BlockDeleteError"""
-        pass
-
-    @abc.abstractmethod
-    def download_file(self, key: str, filename: str) -> None:
-        pass
-
-    @abc.abstractmethod
-    def upload_file(self, key: str, filename: str) -> None:
-        pass
-
-    @abc.abstractmethod
-    def delete_file(self, key: str) -> None:
-        pass
-
-
-class UnavailableSessionRecordingV2ObjectStorage(SessionRecordingV2ObjectStorageBase):
-    def read_bytes(self, key: str, first_byte: int, last_byte: int) -> bytes | None:
-        return None
-
-    def read_all_bytes(self, key: str) -> bytes | None:
-        pass
-
-    def write(self, key: str, data: bytes) -> None:
-        pass
-
-    def is_enabled(self) -> bool:
-        return False
-
-    def fetch_file(self, blob_key: str) -> str:
-        raise FileFetchError("Storage not available")
-
-    def fetch_file_bytes(self, blob_key: str) -> bytes:
-        raise FileFetchError("Storage not available")
-
-    def fetch_block(self, block_url: str) -> str:
-        raise BlockFetchError("Storage not available")
-
-    def fetch_block_bytes(self, block_url: str) -> bytes:
-        raise BlockFetchError("Storage not available")
-
-    def delete_block(self, block_url: str) -> None:
-        raise BlockDeleteError("Storage not available")
-
-    def store_lts_recording(self, recording_id: str, recording_data: str) -> tuple[Optional[str], Optional[str]]:
-        return None, "Storage not available"
-
-    def is_lts_enabled(self) -> bool:
-        return False
-
-    def download_file(self, key: str, filename: str) -> None:
-        raise FileDownloadError("Storage not available")
-
-    def upload_file(self, key: str, filename: str) -> None:
-        raise FileUploadError("Storage not available")
-
-    def delete_file(self, key: str) -> None:
-        raise FileDeleteError("Storage not available")
-
-
-class SessionRecordingV2ObjectStorage(SessionRecordingV2ObjectStorageBase):
-    def __init__(self, aws_client, bucket: str) -> None:
-        self.aws_client = aws_client
-        self.bucket = bucket
-
-    def read_bytes(self, key: str, first_byte: int, last_byte: int) -> bytes | None:
-        s3_response = {}
-        try:
-            kwargs = {
-                "Bucket": self.bucket,
-                "Key": key,
-                "Range": f"bytes={first_byte}-{last_byte}",
-            }
-            s3_response = self.aws_client.get_object(**kwargs)
-            return s3_response["Body"].read()
-        except Exception as e:
-            logger.exception(
-                "recording_block_storage.read_bytes_failed",
-                bucket=self.bucket,
-                file_name=key,
-                error=e,
-                exc_info=False,
-                s3_response=s3_response,
-            )
-            return None
-
-    def read_all_bytes(self, key: str) -> bytes | None:
-        s3_response = {}
-        try:
-            kwargs = {
-                "Bucket": self.bucket,
-                "Key": key,
-            }
-            s3_response = self.aws_client.get_object(**kwargs)
-            return s3_response["Body"].read()
-        except Exception as e:
-            logger.exception(
-                "recording_block_storage.read_all_bytes_failed",
-                bucket=self.bucket,
-                file_name=key,
-                error=e,
-                exc_info=False,
-                s3_response=s3_response,
-            )
-            return None
-
-    def write(self, key: str, data: bytes) -> None:
-        s3_response = {}
-        try:
-            s3_response = self.aws_client.put_object(
-                Bucket=self.bucket,
-                Key=key,
-                Body=data,
-            )
-        except Exception as e:
-            logger.exception(
-                "recording_block_storage.write_failed",
-                bucket=self.bucket,
-                file_name=key,
-                error=e,
-                exc_info=False,
-                s3_response=s3_response,
-            )
-            raise Exception("Failed to write recording data") from e
-
-    def is_enabled(self) -> bool:
-        return True
-
-    def fetch_file(self, blob_key: str) -> str:
-        try:
-            kwargs = {
-                "Bucket": self.bucket,
-                "Key": blob_key,
-            }
-            s3_response = self.aws_client.get_object(**kwargs)
-            file_body = s3_response["Body"].read()
-
-            return snappy.decompress(file_body).decode("utf-8")
-
-        except Exception as e:
-            posthoganalytics.tag("bucket", self.bucket)
-            logger.exception(
-                "recording_block_storage.fetch_file_failed",
-                bucket=self.bucket,
-                file_name=blob_key,
-                error=e,
-                exc_info=False,
-                s3_response=s3_response,
-            )
-            raise FileFetchError(f"Failed to read and decompress file: {str(e)}")
-
-    def fetch_file_bytes(self, blob_key: str) -> bytes:
-        try:
-            kwargs = {
-                "Bucket": self.bucket,
-                "Key": blob_key,
-            }
-            s3_response = self.aws_client.get_object(**kwargs)
-            return s3_response["Body"].read()
-        except Exception as e:
-            posthoganalytics.tag("bucket", self.bucket)
-            logger.exception(
-                "recording_block_storage.fetch_file_bytes_failed",
-                bucket=self.bucket,
-                file_name=blob_key,
-                error=e,
-                exc_info=False,
-            )
-            raise FileFetchError(f"Failed to read compressed file: {str(e)}")
-
-    def _fetch_compressed_block(self, block_url: str) -> bytes:
-        """Internal method to fetch and validate compressed block"""
-        # Parse URL and extract key and byte range
-        parsed_url = urlparse(block_url)
-        key = parsed_url.path.lstrip("/")
-        query_params = parse_qs(parsed_url.query)
-        byte_range = query_params.get("range", [""])[0].replace("bytes=", "")
-        start_byte, end_byte = map(int, byte_range.split("-")) if "-" in byte_range else (None, None)
-
-        if start_byte is None or end_byte is None:
-            raise BlockFetchError("Invalid byte range in block URL")
-
-        expected_length = end_byte - start_byte + 1
-        compressed_block = self.read_bytes(key, first_byte=start_byte, last_byte=end_byte)
-
-        if not compressed_block:
-            raise BlockFetchError("Block content not found")
-
-        if len(compressed_block) != expected_length:
-            raise BlockFetchError(
-                f"Unexpected data length. Expected {expected_length} bytes, got {len(compressed_block)} bytes"
-            )
-
-        return compressed_block
-
-    def fetch_block(self, block_url: str, session_id: str | None = None, team_id: int | None = None) -> str:
-        # session_id and team_id are accepted for interface compatibility but not used for S3 storage
-        try:
-            compressed_block = self._fetch_compressed_block(block_url)
-            decompressed_block = snappy.decompress(compressed_block).decode("utf-8")
-            # Strip any trailing newlines
-            decompressed_block = decompressed_block.rstrip("\n")
-            return decompressed_block
-
-        except BlockFetchError:
-            raise
-        except Exception as e:
-            logger.exception(
-                "recording_block_storage.fetch_block_failed",
-                bucket=self.bucket,
-                block_url=block_url,
-                error=e,
-                exc_info=False,
-            )
-            raise BlockFetchError(f"Failed to read and decompress block: {str(e)}")
-
-    def fetch_block_bytes(self, block_url: str, session_id: str | None = None, team_id: int | None = None) -> bytes:
-        # session_id and team_id are accepted for interface compatibility but not used for S3 storage
-        try:
-            return self._fetch_compressed_block(block_url)
-        except BlockFetchError:
-            raise
-        except Exception as e:
-            logger.exception(
-                "recording_block_storage.fetch_block_bytes_failed",
-                bucket=self.bucket,
-                block_url=block_url,
-                error=e,
-                exc_info=False,
-            )
-            raise BlockFetchError(f"Failed to read compressed block: {str(e)}")
-
-    def delete_block(self, block_url: str) -> None:
-        try:
-            # Parse URL and extract key and byte range
-            parsed_url = urlparse(block_url)
-            key = parsed_url.path.lstrip("/")
-            query_params = parse_qs(parsed_url.query)
-            byte_range = query_params.get("range", [""])[0].replace("bytes=", "")
-            start_byte, end_byte = map(int, byte_range.split("-")) if "-" in byte_range else (None, None)
-
-            if start_byte is None or end_byte is None:
-                raise BlockDeleteError("Invalid byte range in block URL")
-
-            expected_length = end_byte - start_byte + 1
-
-            file_bytes = self.read_all_bytes(key)
-
-            if file_bytes is None:
-                raise BlockDeleteError(f"Failed to read file {key}")
-
-            new_file_bytes = bytes(
-                bytearray(file_bytes[:start_byte]) + bytearray(expected_length) + bytearray(file_bytes[end_byte + 1 :])
-            )
-
-            self.write(key, new_file_bytes)
-        except BlockDeleteError:
-            raise
-        except Exception as e:
-            logger.exception(
-                "recording_block_storage.delete_block_failed",
-                bucket=self.bucket,
-                block_url=block_url,
-                error=e,
-                exc_info=False,
-            )
-            raise BlockDeleteError(f"Failed to delete block: {str(e)}")
-
-    def download_file(self, key: str, filename: str) -> None:
-        try:
-            self.aws_client.download_file(
-                Bucket=self.bucket,
-                Key=key,
-                Filename=filename,
-            )
-        except Exception as e:
-            logger.exception(
-                "recording_block_storage.download_file_failed",
-                bucket=self.bucket,
-                key=key,
-                error=e,
-                exc_info=False,
-            )
-            raise FileDownloadError(f"Failed to download file: {str(e)}")
-
-    def upload_file(self, key: str, filename: str) -> None:
-        try:
-            self.aws_client.upload_file(
-                Bucket=self.bucket,
-                Key=key,
-                Filename=filename,
-            )
-        except Exception as e:
-            logger.exception(
-                "recording_block_storage.upload_file_failed",
-                bucket=self.bucket,
-                key=key,
-                error=e,
-                exc_info=False,
-            )
-            raise FileUploadError(f"Failed to upload file: {str(e)}")
-
-    def delete_file(self, key: str) -> None:
-        try:
-            self.aws_client.delete_object(
-                Bucket=self.bucket,
-                Key=key,
-            )
-        except Exception as e:
-            logger.exception(
-                "recording_block_storage.delete_file_failed",
-                bucket=self.bucket,
-                key=key,
-                error=e,
-                exc_info=False,
-            )
-            raise FileDeleteError(f"Failed to delete file: {str(e)}")
-
-
-class AsyncSessionRecordingV2ObjectStorage:
-    def __init__(self, aws_client, bucket: str) -> None:
-        self.aws_client = aws_client
-        self.bucket = bucket
-
-    async def read_bytes(self, key: str, first_byte: int, last_byte: int) -> bytes | None:
-        s3_response = {}
-        try:
-            kwargs = {
-                "Bucket": self.bucket,
-                "Key": key,
-                "Range": f"bytes={first_byte}-{last_byte}",
-            }
-            s3_response = await self.aws_client.get_object(**kwargs)
-            return await s3_response["Body"].read()
-        except Exception as e:
-            logger.exception(
-                "async_recording_block_storage.read_bytes_failed",
-                bucket=self.bucket,
-                file_name=key,
-                error=e,
-                exc_info=False,
-                s3_response=s3_response,
-            )
-            return None
-
-    async def read_all_bytes(self, key: str) -> bytes | None:
-        s3_response = {}
-        try:
-            kwargs = {
-                "Bucket": self.bucket,
-                "Key": key,
-            }
-            s3_response = await self.aws_client.get_object(**kwargs)
-            return await s3_response["Body"].read()
-        except Exception as e:
-            logger.exception(
-                "async_recording_block_storage.read_all_bytes_failed",
-                bucket=self.bucket,
-                file_name=key,
-                error=e,
-                exc_info=False,
-                s3_response=s3_response,
-            )
-            return None
-
-    async def write(self, key: str, data: bytes) -> None:
-        s3_response = {}
-        try:
-            s3_response = await self.aws_client.put_object(
-                Bucket=self.bucket,
-                Key=key,
-                Body=data,
-            )
-        except Exception as e:
-            logger.exception(
-                "async_recording_block_storage.write_failed",
-                bucket=self.bucket,
-                file_name=key,
-                error=e,
-                exc_info=False,
-                s3_response=s3_response,
-            )
-            raise Exception("Failed to write recording data") from e
-
-    def is_enabled(self) -> bool:
-        return True
-
-    async def fetch_file(self, blob_key: str) -> str:
-        try:
-            kwargs = {
-                "Bucket": self.bucket,
-                "Key": blob_key,
-            }
-            s3_response = await self.aws_client.get_object(**kwargs)
-            file_body = await s3_response["Body"].read()
-
-            return snappy.decompress(file_body).decode("utf-8")
-        except Exception as e:
-            posthoganalytics.tag("bucket", self.bucket)
-            logger.exception(
-                "async_recording_block_storage.fetch_file_failed",
-                bucket=self.bucket,
-                file_name=blob_key,
-                error=e,
-                exc_info=False,
-                s3_response=s3_response,
-            )
-            raise FileFetchError(f"Failed to read and decompress file: {str(e)}")
-
-    async def fetch_file_bytes(self, blob_key: str) -> bytes:
-        try:
-            kwargs = {
-                "Bucket": self.bucket,
-                "Key": blob_key,
-            }
-            s3_response = await self.aws_client.get_object(**kwargs)
-            return await s3_response["Body"].read()
-        except Exception as e:
-            posthoganalytics.tag("bucket", self.bucket)
-            logger.exception(
-                "async_recording_block_storage.fetch_file_bytes_failed",
-                bucket=self.bucket,
-                file_name=blob_key,
-                error=e,
-                exc_info=False,
-            )
-            raise FileFetchError(f"Failed to read compressed file: {str(e)}")
-
-    async def _fetch_compressed_block(self, block_url: str) -> bytes:
-        """Internal method to fetch and validate compressed block"""
-        # Parse URL and extract key and byte range
-        parsed_url = urlparse(block_url)
-        key = parsed_url.path.lstrip("/")
-        query_params = parse_qs(parsed_url.query)
-        byte_range = query_params.get("range", [""])[0].replace("bytes=", "")
-        start_byte, end_byte = map(int, byte_range.split("-")) if "-" in byte_range else (None, None)
-
-        if start_byte is None or end_byte is None:
-            raise BlockFetchError("Invalid byte range in block URL")
-
-        expected_length = end_byte - start_byte + 1
-        compressed_block = await self.read_bytes(key, first_byte=start_byte, last_byte=end_byte)
-
-        if not compressed_block:
-            raise BlockFetchError("Block content not found")
-
-        if len(compressed_block) != expected_length:
-            raise BlockFetchError(
-                f"Unexpected data length. Expected {expected_length} bytes, got {len(compressed_block)} bytes"
-            )
-
-        return compressed_block
-
-    async def fetch_block(self, block_url: str, session_id: str | None = None, team_id: int | None = None) -> str:
-        # session_id and team_id are accepted for interface compatibility but not used for S3 storage
-        try:
-            compressed_block = await self._fetch_compressed_block(block_url)
-            decompressed_block = snappy.decompress(compressed_block).decode("utf-8")
-            # Strip any trailing newlines
-            decompressed_block = decompressed_block.rstrip("\n")
-            return decompressed_block
-
-        except BlockFetchError:
-            raise
-        except Exception as e:
-            logger.exception(
-                "async_recording_block_storage.fetch_block_failed",
-                bucket=self.bucket,
-                block_url=block_url,
-                error=e,
-                exc_info=False,
-            )
-            raise BlockFetchError(f"Failed to read and decompress block: {str(e)}")
-
-    async def fetch_block_bytes(
-        self, block_url: str, session_id: str | None = None, team_id: int | None = None
-    ) -> bytes:
-        # session_id and team_id are accepted for interface compatibility but not used for S3 storage
-        try:
-            return await self._fetch_compressed_block(block_url)
-        except BlockFetchError:
-            raise
-        except Exception as e:
-            logger.exception(
-                "async_recording_block_storage.fetch_block_bytes_failed",
-                bucket=self.bucket,
-                block_url=block_url,
-                error=e,
-                exc_info=False,
-            )
-            raise BlockFetchError(f"Failed to read compressed block: {str(e)}")
-
-    async def delete_block(self, block_url: str) -> None:
-        try:
-            # Parse URL and extract key and byte range
-            parsed_url = urlparse(block_url)
-            key = parsed_url.path.lstrip("/")
-            query_params = parse_qs(parsed_url.query)
-            byte_range = query_params.get("range", [""])[0].replace("bytes=", "")
-            start_byte, end_byte = map(int, byte_range.split("-")) if "-" in byte_range else (None, None)
-
-            if start_byte is None or end_byte is None:
-                raise BlockDeleteError("Invalid byte range in block URL")
-
-            expected_length = end_byte - start_byte + 1
-
-            file_bytes = await self.read_all_bytes(key)
-
-            if file_bytes is None:
-                raise BlockDeleteError(f"Failed to read file {key}")
-
-            new_file_bytes = bytes(
-                bytearray(file_bytes[:start_byte]) + bytearray(expected_length) + bytearray(file_bytes[end_byte + 1 :])
-            )
-
-            await self.write(key, new_file_bytes)
-        except BlockDeleteError:
-            raise
-        except Exception as e:
-            logger.exception(
-                "async_recording_block_storage.delete_block_failed",
-                bucket=self.bucket,
-                block_url=block_url,
-                error=e,
-                exc_info=False,
-            )
-            raise BlockDeleteError(f"Failed to delete block: {str(e)}")
-
-    async def download_file(self, key: str, filename: str) -> None:
-        try:
-            await self.aws_client.download_file(
-                Bucket=self.bucket,
-                Key=key,
-                Filename=filename,
-            )
-        except Exception as e:
-            logger.exception(
-                "async_recording_block_storage.download_file_failed",
-                bucket=self.bucket,
-                key=key,
-                error=e,
-                exc_info=False,
-            )
-            raise FileDownloadError(f"Failed to download file: {str(e)}")
-
-    async def upload_file(self, key: str, filename: str) -> None:
-        try:
-            await self.aws_client.upload_file(
-                Bucket=self.bucket,
-                Key=key,
-                Filename=filename,
-            )
-        except Exception as e:
-            logger.exception(
-                "async_recording_block_storage.upload_file_failed",
-                bucket=self.bucket,
-                key=key,
-                error=e,
-                exc_info=False,
-            )
-            raise FileUploadError(f"Failed to upload file: {str(e)}")
-
-    async def delete_file(self, key: str) -> None:
-        try:
-            await self.aws_client.delete_object(
-                Bucket=self.bucket,
-                Key=key,
-            )
-        except Exception as e:
-            logger.exception(
-                "async_recording_block_storage.delete_file_failed",
-                bucket=self.bucket,
-                key=key,
-                error=e,
-                exc_info=False,
-            )
-            raise FileDeleteError(f"Failed to delete file: {str(e)}")
-
-
-_client: SessionRecordingV2ObjectStorageBase = UnavailableSessionRecordingV2ObjectStorage()
-
-
-def client() -> SessionRecordingV2ObjectStorageBase:
-    global _client
-
-    required_settings = [
-        settings.SESSION_RECORDING_V2_S3_ENDPOINT,
-        settings.SESSION_RECORDING_V2_S3_REGION,
-        settings.SESSION_RECORDING_V2_S3_BUCKET,
-    ]
-
-    if not all(required_settings):
-        _client = UnavailableSessionRecordingV2ObjectStorage()
-    elif isinstance(_client, UnavailableSessionRecordingV2ObjectStorage):
-        _client = SessionRecordingV2ObjectStorage(
-            boto3_client(
-                "s3",
-                endpoint_url=settings.SESSION_RECORDING_V2_S3_ENDPOINT,
-                aws_access_key_id=settings.SESSION_RECORDING_V2_S3_ACCESS_KEY_ID,
-                aws_secret_access_key=settings.SESSION_RECORDING_V2_S3_SECRET_ACCESS_KEY,
-                config=Config(
-                    signature_version="s3v4",
-                    connect_timeout=1,
-                    retries={"max_attempts": 1},
-                ),
-                region_name=settings.SESSION_RECORDING_V2_S3_REGION,
-            ),
-            bucket=settings.SESSION_RECORDING_V2_S3_BUCKET,
-        )
-
-    return _client
-
-
-@asynccontextmanager
-async def async_client() -> AsyncIterator[AsyncSessionRecordingV2ObjectStorage]:
-    required_settings = [
-        settings.SESSION_RECORDING_V2_S3_ENDPOINT,
-        settings.SESSION_RECORDING_V2_S3_REGION,
-        settings.SESSION_RECORDING_V2_S3_BUCKET,
-    ]
-
-    if not all(required_settings):
-        raise RuntimeError("Missing required settings for object storage client")
-    else:
-        session = aioboto3.Session()
-        async with session.client(  # type: ignore[call-overload]
-            "s3",
-            endpoint_url=settings.SESSION_RECORDING_V2_S3_ENDPOINT,
-            aws_access_key_id=settings.SESSION_RECORDING_V2_S3_ACCESS_KEY_ID,
-            aws_secret_access_key=settings.SESSION_RECORDING_V2_S3_SECRET_ACCESS_KEY,
-            config=Config(
-                signature_version="s3v4",
-                connect_timeout=1,
-                retries={"max_attempts": 1},
-            ),
-            region_name=settings.SESSION_RECORDING_V2_S3_REGION,
-        ) as client:
-            yield AsyncSessionRecordingV2ObjectStorage(
-                client,
-                bucket=settings.SESSION_RECORDING_V2_S3_BUCKET,
-            )
-
-
-class EncryptedBlockStorage:
-    """
-    Async client for fetching recording blocks via the Recording API.
-
-    The Recording API handles decryption transparently - encrypted sessions are
-    decrypted using KMS, while unencrypted sessions pass through unchanged.
-    This allows gradual migration from direct S3 reads to API-based reads.
-    """
-
-    def __init__(self, session: aiohttp.ClientSession, base_url: str):
-        self.session = session
-        self.base_url = base_url.rstrip("/")
-
-    def _parse_block_url(self, block_url: str) -> tuple[str, int, int]:
-        """
-        Parse a block URL to extract the key, start byte, and end byte.
-
-        The block_url is in the format: s3://bucket/key?range=bytes=start-end
-        Returns a tuple of (key, start, end).
-        """
-        parsed = urlparse(block_url)
-        key = parsed.path.lstrip("/")
-
-        match = re.match(r"^range=bytes=(\d+)-(\d+)$", parsed.query)
-        if not match:
-            raise ValueError(f"Invalid range format: {parsed.query}")
-
-        return key, int(match.group(1)), int(match.group(2))
-
-    async def fetch_block_bytes(self, block_url: str, session_id: str, team_id: int) -> bytes:
-        """
-        Fetch a recording block via the Recording API.
-
-        Returns the decrypted but still snappy-compressed block bytes.
-
-        Raises:
-            RecordingDeletedError: If the recording has been deleted (crypto shredding).
-            RecordingApiFetchError: If the block is not found or other fetch errors occur.
-        """
-        key, start, end = self._parse_block_url(block_url)
-        url = f"{self.base_url}/api/projects/{team_id}/recordings/{session_id}/block"
-
-        try:
-            async with self.session.get(url, params={"key": key, "start": start, "end": end}) as response:
-                if response.status == 404:
-                    raise RecordingApiFetchError("Block not found")
-                if response.status == 410:
-                    data = await response.json()
-                    deleted_at = data.get("deleted_at")
-                    logger.info(
-                        "encrypted_block_storage.recording_deleted",
-                        session_id=session_id,
-                        team_id=team_id,
-                        deleted_at=deleted_at,
-                    )
-                    raise RecordingDeletedError("Recording has been deleted", deleted_at=deleted_at)
-                response.raise_for_status()
-                return await response.read()
-        except (RecordingDeletedError, RecordingApiFetchError):
-            raise
-        except aiohttp.ClientError as e:
-            logger.exception(
-                "encrypted_block_storage.fetch_block_bytes_failed",
-                url=url,
-                session_id=session_id,
-                team_id=team_id,
-                error=str(e),
-                exc_info=False,
-            )
-            raise RecordingApiFetchError(f"Failed to fetch block from Recording API: {str(e)}")
-
-    async def fetch_block(self, block_url: str, session_id: str, team_id: int) -> str:
-        """
-        Fetch and decompress a recording block via the Recording API.
-
-        Returns the decrypted and decompressed block content as a string.
-
-        Raises:
-            RecordingDeletedError: If the recording has been deleted (crypto shredding).
-            RecordingApiFetchError: If the block is not found or other fetch errors occur.
-        """
-        try:
-            compressed_block = await self.fetch_block_bytes(block_url, session_id, team_id)
-            decompressed_block = snappy.decompress(compressed_block).decode("utf-8")
-            return decompressed_block.rstrip("\n")
-        except (RecordingDeletedError, RecordingApiFetchError):
-            raise
-        except Exception as e:
-            logger.exception(
-                "encrypted_block_storage.fetch_block_failed",
-                session_id=session_id,
-                team_id=team_id,
-                error=str(e),
-                exc_info=False,
-            )
-            raise RecordingApiFetchError(f"Failed to decompress block: {str(e)}")
-
-    async def delete_recording(self, session_id: str, team_id: int) -> bool:
-        """
-        Delete a recording's encryption key via the Recording API (crypto shredding).
-
-        Returns True if the key was deleted.
-
-        Raises:
-            RecordingDeletedError: If the recording has already been deleted.
-            RecordingApiFetchError: If the recording key is not found or other errors occur.
-        """
-        url = f"{self.base_url}/api/projects/{team_id}/recordings/{session_id}"
-
-        try:
-            async with self.session.delete(url) as response:
-                if response.status == 404:
-                    raise RecordingApiFetchError("Recording key not found")
-                if response.status == 410:
-                    data = await response.json()
-                    deleted_at = data.get("deleted_at")
-                    logger.info(
-                        "encrypted_block_storage.recording_already_deleted",
-                        session_id=session_id,
-                        team_id=team_id,
-                        deleted_at=deleted_at,
-                    )
-                    raise RecordingDeletedError("Recording has already been deleted", deleted_at=deleted_at)
-                response.raise_for_status()
-                return True
-        except (RecordingDeletedError, RecordingApiFetchError):
-            raise
-        except aiohttp.ClientError as e:
-            logger.exception(
-                "encrypted_block_storage.delete_recording_failed",
-                url=url,
-                session_id=session_id,
-                team_id=team_id,
-                error=str(e),
-                exc_info=False,
-            )
-            raise RecordingApiFetchError(f"Failed to delete recording: {str(e)}")
-
-
-@asynccontextmanager
-async def encrypted_block_storage() -> AsyncIterator[EncryptedBlockStorage]:
-    """
-    Async context manager for creating a EncryptedBlockStorage.
-
-    Usage:
-        async with encrypted_block_storage() as client:
-            content = await client.fetch_block(block_url, session_id, team_id)
-    """
-    if not settings.RECORDING_API_URL:
-        raise RuntimeError("RECORDING_API_URL is not configured")
-
-    timeout = aiohttp.ClientTimeout(total=30, connect=5)
-    async with aiohttp.ClientSession(timeout=timeout) as session:
-        yield EncryptedBlockStorage(session, settings.RECORDING_API_URL)
