+++ posthog/storage/test/recordings/__init__.py
@@ -1,0 +1,1 @@
+# Tests for session recording storage modules
